# openapi.yaml
openapi: 3.0.3
info:
  title: SoroSim Backend API
  description: |
    Soroban Contract Simulation & Dry-Run Sandbox Backend - A REST API for simulating 
    Soroban smart contract invocations with mock ledger state.
    
    ## Features
    - Simulate Soroban contract invocations without testnet
    - Mock ledger state management
    - Session-based simulation history
    - Network configuration management
    - State diff tracking and analysis
    - Event extraction and filtering
    - Comprehensive reporting
    
  version: 1.0.0
  contact:
    name: SoroSim
    url: https://github.com/YOUR_USERNAME/SoroSim-backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.sorosim.example.com
    description: Production server (example)
...
paths:
...
  /api/simulate:
    post:
      tags:
        - Simulation
      summary: Simulate contract invocation
      description: Simulate a Soroban contract function call without submitting to the network
      parameters:
        - name: networkId
          in: query
          description: Network ID to use for simulation
          schema:
            type: string
            example: testnet
        - name: sessionId
          in: query
          description: Session ID to log invocation to
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contractId
                - method
              properties:
                contractId:
                  type: string
                  description: Contract address
                  example: CA3D5KRYM6CB7OWQ6TWYRR3Z4T7GNZLKERYNZGGA5SOAOPIFY6YQGAXE
                method:
                  type: string
                  description: Contract function name
                  example: increment
                args:
                  type: array
                  description: Function arguments
                  items:
                    type: object
                  default: []
                source:
                  type: string
                  description: Source account public key
                  example: GABC123...
                fee:
                  type: string
                  description: Transaction fee in stroops
                  example: "1000"
      responses:
        '200':
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimulationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/sessions:
    get:
      tags:
        - Sessions
      summary: Get all sessions
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Session'
    
    post:
      tags:
        - Sessions
      summary: Create new session
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                networkPassphrase:
                  type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    $ref: '#/components/schemas/Session'
...
  /api/networks:
    get:
      tags:
        - Networks
      summary: Get all networks
      responses:
        '200':
          description: List of networks
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Network'
    
    post:
      tags:
        - Networks
      summary: Add custom network
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - id
                - name
                - rpcUrl
                - networkPassphrase
              properties:
                id:
                  type: string
                name:
                  type: string
                rpcUrl:
                  type: string
                networkPassphrase:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Network added
...
components:
  schemas:
    SimulationResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            success:
              type: boolean
            result:
              type: object
            events:
              type: array
              items:
                type: object
            stateDiff:
              type: object
        requestId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time
        duration:
          type: integer
          description: Duration in milliseconds
        sessionId:
          type: string
          format: uuid
...
    Session:
      type: object
      properties:
        sessionId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, idle, closed]
        invocationCount:
          type: integer
        metadata:
          type: object
    Network:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [testnet, futurenet, mainnet, custom]
        rpcUrl:
          type: string
        networkPassphrase:
          type: string
        description:
          type: string
        isDefault:
          type: boolean
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Error message
        details:
          type: object
          description: Additional error details